safari-534.53-branch/LayoutTests/fast/js/closure-inside-extra-arg-call-expected.txt

This test checks that activation objects for functions called with too many arguments are created properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS c1() is "xyxy"
PASS c2() is "xyxy"
PASS successfullyParsed is true

TEST COMPLETE
